Updated tests, still more to be added
