+++ force_wfmanager/plugins/ui_notification/tests/test_ui_notification_hooks_manager.py
@@ -2,6 +2,8 @@ import unittest
 
 from pyface.tasks.task import Task
 
+from envisage.ui.tasks.api import TasksApplication
+
 from force_bdss.api import FactoryRegistryPlugin, Workflow
 from force_wfmanager.plugins.ui_notification.ui_notification_model import \
     UINotificationModel
@@ -24,18 +26,19 @@ class TestUINotificationHooksManager(unittest.TestCase):
 
     def test_before_and_after_execution(self):
         manager = self.factory.create_ui_hooks_manager()
+        workflow = Workflow()
 
         mock_task = mock.Mock(spec=Task)
+        mock_task.app = mock.Mock(spec=TasksApplication)
+        mock_task.app.workflow_m = workflow
         mock_registry = mock.Mock(spec=FactoryRegistryPlugin)
-        mock_task.factory_registry = mock_registry
+        mock_task.app.factory_registry = mock_registry
         mock_server = mock.Mock(spec=ZMQServer)
         mock_server.ports = (54537, 54538)
-        mock_task.zmq_server = mock_server
+        mock_task.app.zmq_server = mock_server
         mock_registry.notification_listener_factory_by_id.return_value \
             = self.nl_factory
 
-        workflow = Workflow()
-        mock_task.workflow_m = workflow
         manager.before_execution(mock_task)
 
         model = workflow.notification_listeners[0]
