+++ force_wfmanager/tests/test_wfmanager_plugin.py
@@ -4,7 +4,6 @@ from unittest import mock
 from envisage.api import Application
 
 from force_wfmanager.wfmanager_plugin import WfManagerPlugin
-from force_wfmanager.wfmanager_task import WfManagerTask
 
 
 class TestWfManagerPlugin(unittest.TestCase):
@@ -17,32 +16,27 @@ class TestWfManagerPlugin(unittest.TestCase):
         self.wfmanager_plugin_file.application.get_plugin = lambda pl_id: None
 
     def test_init(self):
-        self.assertEqual(len(self.wfmanager_plugin.tasks), 1)
+        self.assertEqual(len(self.wfmanager_plugin.tasks), 2)
         self.assertEqual(self.wfmanager_plugin.tasks[0].name,
-                         "Workflow Manager")
+                         "Workflow Manager (Setup)")
+        self.assertEqual(self.wfmanager_plugin.tasks[1].name,
+                         "Workflow Manager (Results)")
 
         def mock_wfmanager_task_constructor(*args, **kwargs):
             return
 
         with mock.patch(
                 "force_wfmanager.wfmanager_plugin"
-                ".WfManagerTask") as mock_task:
-            mock_task.side_effect = mock_wfmanager_task_constructor
+                ".WfManagerSetupTask") as mock_setup_task:
+            mock_setup_task.side_effect = mock_wfmanager_task_constructor
 
-            self.wfmanager_plugin._create_task()
-            self.assertTrue(mock_task.called)
-
-    def test_file_load(self):
-
-        def mock_wfmanager_task_constructor(*args, **kwargs):
-            wfm_task = mock.Mock(spec=WfManagerTask)
-            wfm_task.open_workflow_file = lambda: None
-            return mock_task
+            self.wfmanager_plugin._create_setup_task()
+            self.assertTrue(mock_setup_task.called)
 
         with mock.patch(
                 "force_wfmanager.wfmanager_plugin"
-                ".WfManagerTask") as mock_task:
-            mock_task.side_effect = mock_wfmanager_task_constructor
+                ".WfManagerResultsTask") as mock_results_task:
+            mock_results_task.side_effect = mock_wfmanager_task_constructor
 
-            mock_task = self.wfmanager_plugin_file._create_task()
-            self.assertTrue(mock_task.open_workflow_file.called)
+            self.wfmanager_plugin._create_results_task()
+            self.assertTrue(mock_results_task.called)
